upgrade Traccar to v6.8.0
restructured documentation
moved websocket doc into traccar-api/dox
upgraded docker to traccar/traccar:6.8.0-ubuntu
improved openapitools-client with online link to yaml
workaround for inconsistencies Long <> Integer
fixed narrowing cast from Long to Integer value range

+++ bm-traccar/traccar-api-client/src/test/java/bm/traccar/api/BaseIntegrationTest.java
@@ -47,8 +47,9 @@ public abstract class BaseIntegrationTest {
   protected String userMail;
 
   @Autowired protected ApiService api;
+  // protected Api api; // interface only
 
-  int userId, adminId;
+  Long userId, adminId;
 
   @BeforeAll
   public void setup() throws ApiException {
